# Benchmark `WriteBytesAsync` optimization