--- a/Giraffe.sln
+++ b/Giraffe.sln
@@ -32,6 +32,8 @@ Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "_root", "_root", "{FADD8661
 EndProject
 Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "ResponseCachingApp", "samples\ResponseCachingApp\ResponseCachingApp.fsproj", "{FA102AC4-4608-42F9-86C1-1472B416A76E}"
 EndProject
+Project("{F2A71F9B-5D33-465A-A702-920D77279786}") = "Benchmarks", "benchmarks\Benchmarks.fsproj", "{8C9B540A-EF5E-44A7-95B1-63444E326D68}"
+EndProject
 Global
 	GlobalSection(SolutionConfigurationPlatforms) = preSolution
 		Debug|Any CPU = Debug|Any CPU
@@ -90,6 +92,18 @@ Global
 		{FA102AC4-4608-42F9-86C1-1472B416A76E}.Release|x64.Build.0 = Release|Any CPU
 		{FA102AC4-4608-42F9-86C1-1472B416A76E}.Release|x86.ActiveCfg = Release|Any CPU
 		{FA102AC4-4608-42F9-86C1-1472B416A76E}.Release|x86.Build.0 = Release|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Debug|x64.ActiveCfg = Debug|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Debug|x64.Build.0 = Debug|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Debug|x86.ActiveCfg = Debug|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Debug|x86.Build.0 = Debug|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Release|Any CPU.Build.0 = Release|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Release|x64.ActiveCfg = Release|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Release|x64.Build.0 = Release|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Release|x86.ActiveCfg = Release|Any CPU
+		{8C9B540A-EF5E-44A7-95B1-63444E326D68}.Release|x86.Build.0 = Release|Any CPU
 	EndGlobalSection
 	GlobalSection(SolutionProperties) = preSolution
 		HideSolutionNode = FALSE